requirements.txt
# Core dependencies
msal==1.24.1
requests==2.31.0
python-dotenv==1.0.0

# Graph SDK
msgraph-sdk==1.0.0

# JSON/YAML processing
pyyaml==6.0.1
jsonschema==4.19.1
deepdiff==6.5.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utilities
click==8.1.7
colorama==0.4.6
tabulate==0.9.0
